+++ backend/shepherd/src/main/java/com/fisæ/shepherd/api/controller/media/MediaWriteController.java
@@ -1,4 +1,4 @@
-package com.fisæ.shepherd.api.controllers.media;
+package com.fisæ.shepherd.api.controller.media;
 
 import com.fisæ.shepherd.application.media.MediaCommandService;
 import io.swagger.annotations.Api;
